pyproject.toml
[project]
name = "telegram-offers-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
